SVD
Former-commit-id: c3706c1b3cbebf3b07f8f4fc2d916646980c82d4

+++ user/chenyk/svdutil.c
@@ -16,9 +16,7 @@
   along with this program; if not, write to the Free Software
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 */
-
-#include <stdio.h>
-#include <math.h>
+#include <string.h>
 #include <rsf.h>
 #include "svdutil.h"
 
@@ -67,22 +65,22 @@ void ppp(float *a,float *e,float *s,float *v,int m, int n)
 void sss(float fg[2], float cs[2] )
 { 
      float r,d;
-     if ((fabs(fg[0])+fabs(fg[1]))==0.0)  {
+     if ((fabsf(fg[0])+fabsf(fg[1]))==0.0)  {
 	 cs[0]=1.0; cs[1]=0.0; d=0.0;}
      else  {
-	 d=sqrt(fg[0]*fg[0]+fg[1]*fg[1]);
-	 if (fabs(fg[0])>fabs(fg[1])) {
-	     d=fabs(d);
+	 d=sqrtf(fg[0]*fg[0]+fg[1]*fg[1]);
+	 if (fabsf(fg[0])>fabsf(fg[1])) {
+	     d=fabsf(d);
 	     if (fg[0]<0.0) d=-d;
 	 }
-	 if (fabs(fg[1])>=fabs(fg[0])) {
-	     d=fabs(d);
+	 if (fabsf(fg[1])>=fabsf(fg[0])) {
+	     d=fabsf(d);
 	     if (fg[1]<0.0) d=-d;
 	 }
 	 cs[0]=fg[0]/d; cs[1]=fg[1]/d;
      }
     r=1.0;
-    if (fabs(fg[0])>fabs(fg[1])) r=cs[1];
+    if (fabsf(fg[0])>fabsf(fg[1])) r=cs[1];
     else
 	if (cs[0]!=0.0) r=1.0/cs[0];
     fg[0]=d; fg[1]=r;
@@ -108,11 +106,11 @@ int svduov(float *a, float *u,float *o, float *v,double eps,int ka)
                 for (i=kk; i<=m; i++) {
 		    ix=(i-1)*n+kk-1; d=d+a[ix]*a[ix];
 		}
-                s[kk-1]=sqrt(d);
+                s[kk-1]=sqrtf(d);
                 if (s[kk-1]!=0.0) {
 		    ix=(kk-1)*n+kk-1;
                     if (a[ix]!=0.0) {
-			s[kk-1]=fabs(s[kk-1]);
+			s[kk-1]=fabsf(s[kk-1]);
                         if (a[ix]<0.0) s[kk-1]=-s[kk-1];
 		    }
                     for (i=kk; i<=m; i++) {
@@ -152,10 +150,10 @@ int svduov(float *a, float *u,float *o, float *v,double eps,int ka)
 		d=0.0;
                 for (i=kk+1; i<=n; i++)
 		    d=d+e[i-1]*e[i-1];
-                e[kk-1]=sqrt(d);
+                e[kk-1]=sqrtf(d);
                 if (e[kk-1]!=0.0) {
 		    if (e[kk]!=0.0) {
-			e[kk-1]=fabs(e[kk-1]);
+			e[kk-1]=fabsf(e[kk-1]);
                         if (e[kk]<0.0) e[kk-1]=-e[kk-1];
 		    }
                     for (i=kk+1; i<=n; i++)
@@ -264,9 +262,9 @@ int svduov(float *a, float *u,float *o, float *v,double eps,int ka)
 	return(-1);
 	}
         kk=mm-1;
-	while ((kk!=0)&&(fabs(e[kk-1])!=0.0)) {
-	    d=fabs(s[kk-1])+fabs(s[kk]);
-            dd=fabs(e[kk-1]);
+	while ((kk!=0)&&(fabsf(e[kk-1])!=0.0)) {
+	    d=fabsf(s[kk-1])+fabsf(s[kk]);
+            dd=fabsf(e[kk-1]);
             if (dd>eps*d) kk=kk-1;
             else e[kk-1]=0.0;
 	}
@@ -296,24 +294,24 @@ int svduov(float *a, float *u,float *o, float *v,double eps,int ka)
 	}
         else {
 	    ks=mm;
-            while ((ks>kk)&&(fabs(s[ks-1])!=0.0)) {
+            while ((ks>kk)&&(fabsf(s[ks-1])!=0.0)) {
 		d=0.0;
-                if (ks!=mm) d=d+fabs(e[ks-1]);
-                if (ks!=kk+1) d=d+fabs(e[ks-2]);
-                dd=fabs(s[ks-1]);
+                if (ks!=mm) d=d+fabsf(e[ks-1]);
+                if (ks!=kk+1) d=d+fabsf(e[ks-2]);
+                dd=fabsf(s[ks-1]);
                 if (dd>eps*d) ks=ks-1;
                 else s[ks-1]=0.0;
 	    }
             if (ks==kk) {
 		kk=kk+1;
-                d=fabs(s[mm-1]);
-                t=fabs(s[mm-2]);
+                d=fabsf(s[mm-1]);
+                t=fabsf(s[mm-2]);
                 if (t>d) d=t;
-                t=fabs(e[mm-2]);
+                t=fabsf(e[mm-2]);
                 if (t>d) d=t;
-                t=fabs(s[kk-1]);
+                t=fabsf(s[kk-1]);
                 if (t>d) d=t;
-                t=fabs(e[kk-1]);
+                t=fabsf(e[kk-1]);
                 if (t>d) d=t;
                 sm=s[mm-1]/d; sm1=s[mm-2]/d;
                 em1=e[mm-2]/d;
@@ -321,7 +319,7 @@ int svduov(float *a, float *u,float *o, float *v,double eps,int ka)
                 b=((sm1+sm)*(sm1-sm)+em1*em1)/2.0;
                 c=sm*em1; c=c*c; shh=0.0;
                 if ((b!=0.0)||(c!=0.0)) {
-		    shh=sqrt(b*b+c);
+		    shh=sqrtf(b*b+c);
                     if (b<0.0) shh=-shh;
                     shh=c/(b+shh);
 		}
